+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## [1.1.0] - Release
+
+* Added styling of the country selection button.
+* Added support for input formatters.
+* Minor fixes.
+
+
 ## [1.0.1] - Release
 
 * A Basic validator added.
-* Update README.md
+* Updated README.
 
 ## [1.0.0] - Release